Moving along now have added a new package react-navigation to project for easy navigation.
